+++ packages/core/CHANGELOG.md
@@ -11,7 +11,9 @@ Versioning].
 - Fix cases where HTML attributes were not mangled due to whitespace.
 - Fix mangling multiple HTML attributes on a single HTML element.
 - Fix mangling multiple HTML attributes in a single HTML selector in JavaScript.
+- Fix mangling HTML attributes when `>` appears in an attribute value.
 - Support mangling HTML attribute usage (with `attr(data-foo)`) in CSS.
+- Support mangling attribute-value selectors in JavaScript.
 - Improve HTML attribute mangler performance on JavaScript.
 
 ## [0.1.7] - 2021-02-10